Refactor numbering module and add UI panel
Moved numbering logic from module/numbering/util.py to bonsai/core/numbering.py and deleted util.py. Updated imports across the module to use bonsai.core.numbering. Moved all the numbering settings to a separate UI panel (BIM_PT_Numbering), defined in ui.py, which is shown in the Tool panel of the sidebar, in a separate panel. Updated operator logic to use new Numbering class methods. Improved property naming for clarity and consistency.
